UserController - error renaming

+++ Controllers/UserController.cs
@@ -38,7 +38,7 @@ namespace GliwickiDzik.Controllers
             var userForGet = await _repository.GetOneUserAsync(userId);
 
             if (userForGet == null)
-                return BadRequest("The user cannot be found!");
+                return BadRequest("Error: The user cannot be found!");
 
             var userToReturn = _mapper.Map<UserForReturnDTO>(userForGet);
 
@@ -50,8 +50,8 @@ namespace GliwickiDzik.Controllers
         {
             var usersToList = await _repository.GetAllUsersForRecords(userParams);
 
-            if (usersToList == null)
-                return BadRequest("Users cannot be found!");
+            if (usersToList.Count == 0)
+                return NoContent();
             
             var listedUsers = _mapper.Map<IEnumerable<UserForRecordsDTO>>(usersToList);
 
@@ -68,12 +68,15 @@ namespace GliwickiDzik.Controllers
             
             var userForEdit = await _repository.GetOneUserAsync(userId);
 
+            if (userForEdit == null)
+                return BadRequest("Error: The user cannot be found!");
+
             _mapper.Map(userForEditDTO, userForEdit);
 
             if (await _repository.SaveAllUserContent())
                 return NoContent();
             
-            throw new Exception("Error occured while trying to save in database");
+            throw new Exception("Error: Saving edited user to database failed!");
         }
 
         [HttpDelete("DeleteUser")]
@@ -83,13 +86,16 @@ namespace GliwickiDzik.Controllers
                 return Unauthorized();
 
             var userToDelete = await _repository.GetOneUserAsync(userId);
+
+            if (userToDelete == null)
+                return BadRequest("Error: The user cannot be found!");
             
             _repository.Remove(userToDelete);
 
             if (await _repository.SaveAllUserContent())
                 return NoContent();
             
-            throw new Exception("Error occured while trying to save in database");
+            throw new Exception("Error: Removing user from database failed!");
         }
 
         [HttpPost("AddLike/{trainingPlanId}")]
@@ -104,7 +110,7 @@ namespace GliwickiDzik.Controllers
             var trainingPlanToLike = await _trainingRepository.GetOneTrainingPlanAsync(trainingPlanId);
 
             if (trainingPlanToLike == null)
-                return BadRequest("The trainingPlan cannot be found!");
+                return BadRequest("Error: The training plan cannot be found!");
             
             var like = new LikeModel
             {
@@ -119,7 +125,7 @@ namespace GliwickiDzik.Controllers
             if (await _trainingRepository.SaveAllTrainingContent() && await _repository.SaveAllUserContent())
                 return NoContent();
 
-            throw new Exception("Errorc occured while trying save changes to database!");
+            throw new Exception("Error: Saving like to database failed!");
         }
 
         [HttpPost("DeleteLike/{trainingPlanId}")]
@@ -129,12 +135,12 @@ namespace GliwickiDzik.Controllers
                 return Unauthorized();
             
             if (!await _repository.IsLikedAsync(userId, trainingPlanId))
-                return BadRequest("You have to first liked this training plan!");
+                return BadRequest("Error: Like doesn't exist!");
 
             var trainingPlanToDislike = await _trainingRepository.GetOneTrainingPlanAsync(trainingPlanId);
 
             if (trainingPlanToDislike == null)
-                return BadRequest("The trainingPlan cannot be found!");
+                return BadRequest("Error: The training plan cannot be found!");
 
             var likeToRemove = await _repository.GetLikeAsync(userId, trainingPlanId);
 
@@ -145,7 +151,7 @@ namespace GliwickiDzik.Controllers
             if (await _trainingRepository.SaveAllTrainingContent() && await _repository.SaveAllUserContent())
                 return NoContent();
                 
-            throw new Exception("Errorc occured while trying save changes to database!");
+            throw new Exception("Error: Removing like from database failed!");
         }
     }
 }
